#232: moved sanitizer to utility folder
