--- src/integrations/sanitization/definitions/interfaces.ts
@@ -1,5 +1,0 @@
-
-export interface Sanitizer
-{
-    sanitize(input: Record<string, unknown>): Record<string, unknown>;
-}
