--- src/integrations/sanitization/errors/SanitizerError.ts
@@ -1,5 +1,0 @@
-
-export default class SanitizerError extends Error
-{
-
-}
