--- src/integrations/sanitization/errors/UnknownImplementation.ts
@@ -1,10 +1,0 @@
-
-import SanitizerError from './SanitizerError.js';
-
-export default class UnknownImplementation extends SanitizerError
-{
-    constructor(name?: string)
-    {
-        super(`Unknown sanitizer implementation: ${name}`);
-    }
-}
