--- src/integrations/sanitization/implementations/sanitize-html/create.ts
@@ -1,7 +1,0 @@
-
-import SanitizeHtml from './SanitizeHtml.js';
-
-export default function create(): SanitizeHtml
-{
-    return new SanitizeHtml();
-}
