+++ test/integrations/utilities/sanitize.spec.ts
@@ -1,7 +1,7 @@
 
 import { describe, expect, it } from 'vitest';
 
-import sanitizer from '^/integrations/sanitization/module';
+import { sanitize } from '^/integrations/utilities/sanitize';
 
 import { VALUES } from './fixtures';
 
@@ -9,28 +9,28 @@ describe('integrations/sanitization/implementation', () =>
 {
     it('should leave non-string values unchanged', () =>
     {
-        const result = sanitizer.sanitize(VALUES.INPUT.NON_STRING_VALUES);
+        const result = sanitize(VALUES.INPUT.NON_STRING_VALUES);
 
         expect(result).toEqual(VALUES.OUTPUT.NON_STRING_VALUES);
     });
 
     it('should sanitize large HTML strings', () =>
     {
-        const result = sanitizer.sanitize({ large: VALUES.INPUT.LARGE_HTML_STRING });
+        const result = sanitize({ large: VALUES.INPUT.LARGE_HTML_STRING });
 
         expect(result).toEqual({ large: VALUES.OUTPUT.LARGE_HTML_STRING });
     });
 
     it('should NOT sanitize nested objects', () =>
     {
-        const result = sanitizer.sanitize(VALUES.INPUT.NESTED_OBJECT);
+        const result = sanitize(VALUES.INPUT.NESTED_OBJECT);
 
         expect(result).toEqual(VALUES.OUTPUT.NESTED_OBJECT);
     });
 
     it('should NOT sanitize string arrays', () =>
     {
-        const result = sanitizer.sanitize(VALUES.INPUT.OBJECT_WITH_STRING_ARRAY);
+        const result = sanitize(VALUES.INPUT.OBJECT_WITH_STRING_ARRAY);
 
         expect(result).toEqual(VALUES.OUTPUT.OBJECT_WITH_STRING_ARRAY);
     });
